Complete q_insert_head, q_remove_head
In q_insert_head, it needs to check if every malloc work and return the
addresses. It uses strcpy, memset, strcpy to process the strings.
It also needs to check if it insert the first node to assign q->tail to
the new element.

In q_remove_head, it is noticeable that it only copy bufsize-1
characters, plus a null terminator.