Complete q_free, q_size and q_insert_tail
In q_free, I originally free the counter so that q_free couldn't work
in the right way. It needs to declare the counter and assign q->head to
it every time so that it can free the node.

In q_size, I didn't check if q exists. If it doesn't exist, it return 0.

q_insert_tail works like q_insert_head